datalakeservices/requirements.txt
# Required Python Packages

boto3==1.21.25
confluent-kafka==2.3.0
Flask==2.2.5
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
lxml==4.9.3
minio==7.2.0
motor==3.3.2
neo4j==5.7.0
pandas==1.4.1
phonenumbers==8.13.27
plotly==5.18.0
pymilvus==2.3.4
pymongo==4.6.1
PyMuPDF==1.22.5
# psycopg2==2.9.9
protobuf==3.20.0
psycopg2-binary==2.9.9
pytest==7.4.0
PyYAML==6.0.1
ray==2.8.1
redis==4.4.2
scikit-learn==1.3.2
spacy==3.6.1
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-llm==0.4.3
spacy-loggers==1.0.4
spacy-model-manager==0.1.3
typedb-client==2.18.2
typedb-driver==2.26.1
typedb-protocol==2.18.0
typing-inspect==0.8.0
typing_extensions==4.6.3
